Optimize query to fetch relation sizes
Rather than relying on pg_statio_user_tables that is doing
way more than needed here, we just get the user table names
and get their corresponding total relation size. By that we
avoid some joins on pg_index and function calls that we are
not interested in here anyway.